backend_folder/requirements.txt
fastapi
uvicorn
pymongo[srv]
requests
beautifulsoup4
scikit-learn
python-dotenv
dnspython
lxml
playwright